requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
opencv-python
opencv-python-headless
decorator==5.1.1
fastapi==0.115.5
ffmpy==0.4.0
filelock==3.16.1
fonttools==4.55.0
fsspec==2024.10.0
gradio==4.44.1
gradio_client==1.3.0
gradio_image_prompter==0.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
huggingface-hub==0.26.3
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
importlib_resources==6.4.5
Jinja2==3.1.4
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
moviepy==2.0.0
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
numpy-utils==0.1.6
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pillow==10.4.0
proglog==0.1.10
psutil==6.1.0
py-cpuinfo==9.0.0
pydantic==2.10.2
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
ruff==0.8.1
safehttpx==0.1.1
scipy==1.14.1
seaborn==0.13.2
semantic-version==2.10.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.3
tomlkit==0.12.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.0
typer==0.14.0
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.36
ultralytics-thop==2.0.12
urllib3==2.2.3
uvicorn==0.32.1
websockets==12.0